Share mock data for WG views
